Jorge_Alonso-LLM/requirements.txt
flask
flask-cors
groq
python
psycopg2-binary